[awesomescraper] scrape and cache index of awesomelists
This PR adds a script `awesomescraper` that will use the index of
awesomelists on Github, and crawl each list, downloading the raw
markdown into a local directory. You can then pipe each of these raw
awesome lists into our markdown scraper to make into IoK graphs, and
then into `utils/iok.sh put_graph` to upload to IPFS. The graphs
currently are not too great due to the variance in awesomelist styles,
and limitations of our current scraper, but it's good to have all this
stuff cached for experimentation.