[v4] node traversed toggle
Adds actions and reducers for toggling node traversals. Traversals are
stored in a separate Set from the actual IoK graph because we want the
hash of the IoK graph to be consistent, and store all other metadata
separately. In general:

```
{
    graph: <IoK graph, resolves to canonical CID>,
    traversed: <list of traversed node IDs>,
    ...<other metadata>..
}
```

For test, added a button for debugging in `IokText`. We should change
this into a more appropriate UI element.