[awesomescraper][cleaner] start data cleaning for scraped ioks
Until our mdscraper gets good enough, we'll have to manually clean some
IoKs we scrape in automatically. Ideally, all awesomelists follow some
format, such as standadization of header and bullet point use, but
that's hard to enforce. All awesomelists created from IoK -->
awesomelist will be like that, but it's not guaranteed for the other way
around. Now that we have a bunch of awesomelists in this project, we can
clean some up manually. That will:
1. give us more hollistic insight into what awesomelists look like, so
we can design better ways to scrape md content
2. create a curated list of ioks to feature on our webapp
3. have more data to play with, to test edge cases in our tooling, as
well as webapp UI