[awesomescraper][cleaner] clean up more lists and edit iok meta id calc
Clean more awesomelists manually from our cached bunch, and add them to
the README for now. Started a list of small bugs I faced when scraping.
The tool is run manually though, so these bugs are pretty low-priority,
as normal users will not see them, only us devs/operators :)

Also hacked `iok.meta` to use UUID for ID calculation again, instead of
hash. Sometimes the scraper will think there's no text (bug) to hash,
but the hack around that is to just get the ID independently of whatever
is in the node data in the first place. The graph is formatted later on
anyways in the cleanup procedure, as documented in the README.